+++ public/shared/client-logger.js
@@ -246,3 +246,4 @@ function install() {
 }
 
 install();
+